Revert "[WEB-4345] P2P endpoint for voter file upload (#459)"
This reverts commit 0b86266d20cf246808079bbcbd976ff31a7f8c3e.
