+++ src/peerly/peerly.module.ts
@@ -6,20 +6,11 @@ import { PeerlyIdentityService } from './services/peerlyIdentity.service'
 import { PeerlyPhoneListService } from './services/peerlyPhoneList.service'
 import { PeerlyMediaService } from './services/peerlyMedia.service'
 import { PeerlyP2pSmsService } from './services/peerlyP2pSms.service'
-import { P2pPhoneListUploadService } from './services/p2pPhoneListUpload.service'
 import { P2pController } from './p2p.controller'
 import { GoogleModule } from '../vendors/google/google.module'
-import { VoterSharedModule } from '../shared/modules/voterShared.module'
-import { CampaignsModule } from '../campaigns/campaigns.module'
 
 @Module({
-  imports: [
-    HttpModule,
-    JwtModule,
-    GoogleModule,
-    VoterSharedModule,
-    CampaignsModule,
-  ],
+  imports: [HttpModule, JwtModule, GoogleModule],
   controllers: [P2pController],
   providers: [
     PeerlyAuthenticationService,
@@ -27,7 +18,6 @@ import { CampaignsModule } from '../campaigns/campaigns.module'
     PeerlyPhoneListService,
     PeerlyMediaService,
     PeerlyP2pSmsService,
-    P2pPhoneListUploadService,
   ],
   exports: [
     PeerlyAuthenticationService,
