--- src/peerly/schemas/p2pPhoneListResponse.schema.ts
@@ -1,8 +1,0 @@
-import { createZodDto } from 'nestjs-zod'
-import { z } from 'zod'
-
-export class P2pPhoneListResponseSchema extends createZodDto(
-  z.object({
-    token: z.string(),
-  }),
-) {}
