+++ src/peerly/schemas/p2pPhoneListStatus.schema.ts
@@ -1,11 +1,21 @@
 import { createZodDto } from 'nestjs-zod'
 import { z } from 'zod'
 
-const checkPhoneListStatusResponseSchema = z.object({
+const checkPhoneListStatusSuccessResponseSchema = z.object({
+  success: z.literal(true),
   phoneListId: z.number(),
   leadsLoaded: z.number(),
 })
 
-export class CheckPhoneListStatusResponseDto extends createZodDto(
-  checkPhoneListStatusResponseSchema,
+const checkPhoneListStatusFailureResponseSchema = z.object({
+  success: z.literal(false),
+  message: z.string(),
+})
+
+export class CheckPhoneListStatusSuccessResponseDto extends createZodDto(
+  checkPhoneListStatusSuccessResponseSchema,
+) {}
+
+export class CheckPhoneListStatusFailureResponseDto extends createZodDto(
+  checkPhoneListStatusFailureResponseSchema,
 ) {}
