--- src/shared/modules/voterShared.module.ts
@@ -1,15 +1,0 @@
-import { Module } from '@nestjs/common'
-import { VoterDatabaseService } from '../../voters/services/voterDatabase.service'
-import { SlackService } from '../services/slack.service'
-
-/**
- * Shared voter services module to break circular dependencies between voters and peerly modules.
- *
- * This module provides core voter database functionality that can be imported by multiple modules
- * without creating circular dependencies.
- */
-@Module({
-  providers: [VoterDatabaseService, SlackService],
-  exports: [VoterDatabaseService],
-})
-export class VoterSharedModule {}
